refactor: reorganize services and update imports for collections and cards
