+++ frontend/src/repositories/stats/StatsRepositoryFactory.ts
@@ -1,8 +1,7 @@
-import { isGuest } from '@/utils/authUtils'
 import { ApiStatsRepository } from './ApiStatsRepository'
 import { LocalStatsRepository } from './LocalStatsRepository'
 import type { StatsRepository } from './StatsRepository'
 
-export function getStatsRepository(): StatsRepository {
-  return isGuest() ? new LocalStatsRepository() : new ApiStatsRepository()
+export function getStatsRepository(isGuest: boolean): StatsRepository {
+  return isGuest ? new LocalStatsRepository() : new ApiStatsRepository()
 }
