+++ frontend/src/routes/_layout/collections/$collectionId/index.tsx
@@ -6,7 +6,7 @@ import ErrorState from '@/components/commonUI/ErrorState'
 import FloatingActionButton from '@/components/commonUI/FloatingActionButton'
 import ListSkeleton from '@/components/commonUI/ListSkeleton'
 import ScrollableContainer from '@/components/commonUI/ScrollableContainer'
-import { deleteCard, getCards } from '@/services/flashcards/cards'
+import { deleteCard, getCards } from '@/services/cards'
 import { Stack } from '@chakra-ui/react'
 import { useQuery, useQueryClient } from '@tanstack/react-query'
 import { createFileRoute, useNavigate } from '@tanstack/react-router'
