+++ frontend/src/routes/_layout/collections/$collectionId/stats.tsx
@@ -4,7 +4,7 @@ import ErrorState from '@/components/commonUI/ErrorState'
 import LoadingState from '@/components/commonUI/LoadingState'
 import StatsGrids from '@/components/stats/StatsGrids'
 import StatsSummaryGrid from '@/components/stats/StatsSummaryGrid'
-import { getCollectionStats } from '@/services/flashcards/stats'
+import { getCollectionStats } from '@/services/stats'
 import { isGuest } from '@/utils/authUtils'
 import { Box, Container, Heading, Stack, Text } from '@chakra-ui/react'
 import { useQuery } from '@tanstack/react-query'
