+++ frontend/src/services/practiceCards.ts
