+++ frontend/src/services/stats.ts
@@ -1,9 +1,12 @@
 import type { CollectionStats } from '@/client/types.gen'
 import { getStatsRepository } from '@/repositories/stats/StatsRepositoryFactory'
+import { isGuest } from '@/utils/authUtils'
+
+const repo = () => getStatsRepository(isGuest())
 
 export const getCollectionStats = async (
   collectionId: string,
   limit = 30,
 ): Promise<CollectionStats> => {
-  return getStatsRepository().getCollectionStats(collectionId, limit)
+  return repo().getCollectionStats(collectionId, limit)
 }
